+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `lineType` prop
+- `isPlaceholderVisible()` method
 
 ## [0.13.0] - 2019-10-11